MediaSpace/requirements.txt
Django
djangorestframework
django-rest-auth
django-allauth
mysqlclient
